requirement.txt
Django==5.0.1
djangorestframework==3.14.0
django-ratelimit==4.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.2
django-cors-headers
